fix(spa): add participant tile label flickering if webcam is active
the consumer must be aquired in a useEffect, so it will not be available
on the first render. That makes the participant label always render in the
middle of the tile initially, even if the webcam is active. That leads to
the animation of the label executing everytime the tile is recreated.